+++ package.json
@@ -20,7 +20,8 @@
     "node-id3": "^0.1.16",
     "sequelize": "^5.21.11",
     "sharp": "^0.25.3",
-    "shortid": "^2.2.15"
+    "shortid": "^2.2.15",
+    "winston": "^3.2.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.6",
