+++ src/db/index.ts
@@ -1,5 +1,6 @@
 import { Sequelize, Model, DataTypes } from 'sequelize';
 import DotEnv from 'dotenv';
+import Logger from '../util/logger';
 
 DotEnv.config();
 
@@ -9,12 +10,15 @@ const DB = new Sequelize(
   process.env.DB_PASSWORD as string,
   {
     dialect: 'mysql',
+    logging: (msg) => Logger.debug(msg, { label: 'Sequelize' }),
     host: process.env.DB_HOST,
     dialectOptions: {
       socketPath: process.env.DB_SOCKET || '',
     },
   },
 );
+
+DB.sync({ logging: (msg) => Logger.debug(msg, { label: 'Sequelize' }) });
 
 class User extends Model {}
