+++ src/router/index.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+
+const IndexRouter = Router();
+
+IndexRouter.get('/', (_, res) => {
+  res.send('🤔');
+});
+
+export default IndexRouter;
